Submitting Homework 2
I had a few hang-ups with the following in this assignment:

1. displaying the twitter and Facebook icons side by side horizontally
i the footer. I tried to float them and list them in an inline block
and neither was successful. I’m generally feeling shaky in terms of
aligning elements using the float properties.

2. I am confused with using the margins

3. Shaky on how to situate/move divs within the structural elements

4. Unsure as to how I could have written the code to stack the
structural elements so that the entire landing image would be visible,
as in the design jpeg Itho not the file}. It may just be that I should
have used just divs in the body but unsure.

3/5 feeling of confidence on this assignment. I’ll be coming in to
office hours this week!